fix 分布式任务调用任务框架开发 20%

+++ taki-tend-consistency-foramework/tend-cosistency-core/src/main/java/com/taki/model/ConsistencyTaskInstance.java
@@ -60,12 +60,12 @@ public class ConsistencyTaskInstance {
     /**
      * 初始化延迟时间
      */
-    private Integer delayTime;
+    private Long delayTime;
 
     /**
      * 执行时间
      */
-    private Integer executeTime;
+    private Long executeTime;
 
     /**
      * 错误信息
